wip: refactor resource
