+++ libs/frontend/abstract/core/src/domain/action/action.repo.interface.ts
