+++ libs/frontend/domain/builder/src/sections/explorer-pane/BuilderExplorerPane.tsx
@@ -76,11 +76,11 @@ export const BuilderExplorerPane = observer<BuilderMainPaneProps>(
     } = useStore()
 
     const pageBuilderRenderer = builderRenderService.renderers.get(pageId)
+    console.log(pageBuilderRenderer, pageBuilderRenderer?.elementTree)
+
     const root = pageBuilderRenderer?.elementTree.maybeCurrent?.root
     const pageTree = pageBuilderRenderer?.elementTree.maybeCurrent
     const componentId = builderService.activeComponent?.id
-
-    console.log(pageBuilderRenderer?.elementTree)
 
     const componentTree = componentId
       ? builderRenderService.renderers.get(componentId)?.elementTree.current
@@ -276,10 +276,7 @@ export const BuilderExplorerPane = observer<BuilderMainPaneProps>(
         <DeleteFieldModal fieldService={fieldService} />
 
         <CreateActionModal store={appStore} />
-        <UpdateActionModal
-          actionService={actionService}
-          resourceService={resourceService}
-        />
+        <UpdateActionModal />
         <DeleteActionsModal actionService={actionService} />
       </>
     )
