+++ libs/frontend/domain/page/src/view/index.ts
@@ -1,1 +1,2 @@
+export * from './ExplorerPanePage'
 export * from './PageDetailHeader'
