+++ libs/frontend/domain/resource/src/use-cases/get-resource/index.ts
@@ -1,1 +1,1 @@
-export * from './GetResourcesList'
+export * from './ResourcesList'
